Expand signed pin type to 16 bit so we can return 255/-255 (UNUSED_PIN)

+++ MotorDriver.h
@@ -219,7 +219,7 @@ class MotorDriver {
       }
     };
     inline pinpair getSignalPin() { return pinpair(signalPin,signalPin2); };
-    inline int8_t getBrakePinSigned() { return invertBrake ? -brakePin : brakePin; };
+    inline int16_t getBrakePinSigned() { return invertBrake ? -brakePin : brakePin; };
     void setDCSignal(byte speedByte, uint8_t frequency=0);
     void throttleInrush(bool on);
     inline void detachDCSignal() {
